Update cluster.Config defaults to config manager
