+++ pkg/cluster/cluster.go
@@ -1,0 +1,11 @@
+package cluster
+
+func clusterEntry() *ConfigManagerEntry {
+	return &ConfigManagerEntry{
+		Defaulters: []Defaulter{
+			func(c *Config) {
+				c.Cluster.SetDefaults()
+			},
+		},
+	}
+}
